+++ src/despesa_dao.h
@@ -21,7 +21,7 @@ public:
 
     int despesaCount() const;
 
-    qint8 despesaCountUsuario(const QString &email) const;
+    int despesaCountUsuario(const QString &email) const;
 
     std::unique_ptr<std::vector<std::unique_ptr<Despesa>>>
     usuarios(const QString &email = QString()) const;
